added missing examples to phases 1 and 2

+++ phase2/examples/malloc.c
@@ -8,5 +8,5 @@ int main(void)
     int *p;
 
     n = 10;
-    p = malloc(sizeof(*p) * n);
+    p = malloc((int) sizeof(*p) * n);
 }
